backend/requirements.txt
# ITMS Backend Dependencies
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
sqlmodel==0.0.14
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP client for sensor simulation
httpx==0.25.2
aiofiles==23.2.1

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2

# File handling and export
python-multipart==0.0.6
openpyxl==3.1.2
reportlab==4.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Environment and configuration
python-dotenv==1.0.0

# WebSocket support
websockets==12.0

# CORS and middleware
python-cors==1.7.0

# Logging and monitoring
structlog==23.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
